Get rid of KDescendantsProxyModel for Widget3DSubtreeModel
The KDPM was asserting rather randomly on removals. Instead the
Widget3DSubtreeModel now implements both subtree filtering and
flattening of the subtree in single go. This also feels slightly
faster, but more importantly we can now better deal with visibility
changes, new widgets appearing and disappearing etc.

I can reproduce a crash in RemoteModel when certain things change
and new widgets are constructed, but that might just be exposing
a bug in the Widge3DModel (server side).